В класс KuKuMojMalchikTest перемещен исправленный тест из класса KatKuKuMojMalchikTest.  Внесенные исправления в тест: Убраны комментарии, удалены блоки System.out.println, исправлено форматирование. Удален класс KatKuKuMojMalchikTest.
